+++ launcher/src/main/java/com/airbnb/mvrx/launcher/MavericksLauncherMockActivity.kt
@@ -11,7 +11,6 @@ import androidx.core.os.postDelayed
 import androidx.fragment.app.Fragment
 import com.airbnb.mvrx.MavericksBlockExecutions
 import com.airbnb.mvrx.MavericksView
-import com.airbnb.mvrx.MavericksRepositoryConfig
 import com.airbnb.mvrx.launcher.utils.toastLong
 import com.airbnb.mvrx.mocking.MockBehavior
 import com.airbnb.mvrx.mocking.MockedView
