+++ launcher/src/main/java/com/airbnb/mvrx/launcher/ViewModelEnabler.kt
@@ -7,7 +7,6 @@ import androidx.lifecycle.OnLifecycleEvent
 import com.airbnb.mvrx.MavericksBlockExecutions
 import com.airbnb.mvrx.MavericksView
 import com.airbnb.mvrx.MavericksViewModel
-import com.airbnb.mvrx.MavericksRepositoryConfig
 import com.airbnb.mvrx.mocking.MockBehavior
 import com.airbnb.mvrx.mocking.MockableMavericksViewModelConfig
 import com.airbnb.mvrx.mocking.MockedView
