+++ mvrx-core/src/main/java/com/airbnb/mvrx/MavericksRepository.kt
@@ -179,7 +179,7 @@ abstract class MavericksRepository<S : MavericksState>(
         retainValue: KProperty1<S, Async<T>>? = null,
         reducer: S.(Async<T>) -> S
     ): Job {
-        val blockExecutions = onExecute(this@MavericksRepository)
+        val blockExecutions = config.onExecute(this@MavericksRepository)
         if (blockExecutions != MavericksBlockExecutions.No) {
             if (blockExecutions == MavericksBlockExecutions.WithLoading) {
                 setState { reducer(Loading()) }
@@ -219,7 +219,7 @@ abstract class MavericksRepository<S : MavericksState>(
         retainValue: KProperty1<S, Async<T>>? = null,
         reducer: S.(Async<T>) -> S
     ): Job {
-        val blockExecutions = onExecute(this@MavericksRepository)
+        val blockExecutions = config.onExecute(this@MavericksRepository)
         if (blockExecutions != MavericksBlockExecutions.No) {
             if (blockExecutions == MavericksBlockExecutions.WithLoading) {
                 setState { reducer(Loading(value = retainValue?.get(this)?.invoke())) }
@@ -247,7 +247,7 @@ abstract class MavericksRepository<S : MavericksState>(
         dispatcher: CoroutineDispatcher? = null,
         reducer: S.(T) -> S
     ): Job {
-        val blockExecutions = onExecute(this@MavericksRepository)
+        val blockExecutions = config.onExecute(this@MavericksRepository)
         if (blockExecutions != MavericksBlockExecutions.No) {
             // Simulate infinite work
             return coroutineScope.launch { delay(Long.MAX_VALUE) }
@@ -256,28 +256,6 @@ abstract class MavericksRepository<S : MavericksState>(
         return onEach {
             setState { reducer(it) }
         }.launchIn(coroutineScope + (dispatcher ?: EmptyCoroutineContext))
-    }
-
-    /**
-     * Called each time a [execute] function is invoked. This allows
-     * the execute function to be skipped, based on the returned [MavericksBlockExecutions] value.
-     *
-     * This is intended to be used to allow the [MavericksRepository] to be mocked out for testing.
-     * Blocking calls to execute prevents long running asynchronous operations from changing the
-     * state later on when the calls complete.
-     *
-     * Mocking out the state store cannot accomplish this on its own, because in some cases we may
-     * want the state store to initially be mocked, with state changes blocked, but later on we may
-     * want it to allow state changes.
-     *
-     * This prevents the case of an executed async call from modifying state once the state stored
-     * is "enabled", even if the execute was performed when the state store was "disabled" and we
-     * didn't intend to allow operations to change the state.
-     */
-    protected open fun <S : MavericksState> onExecute(
-        repository: MavericksRepository<S>
-    ): MavericksBlockExecutions {
-        return MavericksBlockExecutions.No
     }
 
     /**
