+++ mvrx-mocking/src/main/kotlin/com/airbnb/mvrx/mocking/MockableMavericksViewModelConfig.kt
@@ -4,6 +4,7 @@ import android.content.Context
 import androidx.lifecycle.LifecycleOwner
 import com.airbnb.mvrx.CoroutinesStateStore
 import com.airbnb.mvrx.MavericksBlockExecutions
+import com.airbnb.mvrx.MavericksRepository
 import com.airbnb.mvrx.MavericksViewModel
 import com.airbnb.mvrx.MavericksViewModelConfigFactory
 import com.airbnb.mvrx.MavericksState
@@ -73,11 +74,13 @@ class MockableMavericksViewModelConfig<S : MavericksState>(
 
     fun clearOnExecuteListeners(): Unit = onExecuteListeners.clear()
 
-    override fun <S : MavericksState> onExecute(viewModel: MavericksViewModel<S>): MavericksBlockExecutions {
+    override fun <S : MavericksState> onExecute(repository: MavericksRepository<S>): MavericksBlockExecutions {
         val blockExecutions = currentMockBehavior.blockExecutions
 
-        onExecuteListeners.forEach {
-            it(this, viewModel, blockExecutions)
+        if (repository is MavericksViewModel) {
+            onExecuteListeners.forEach { listener ->
+                listener(this, repository, blockExecutions)
+            }
         }
 
         return blockExecutions
